Support for scala.js
