+++ core/js/src/main/scala/scalaz/zio/internal/OneShot.scala
@@ -26,7 +26,7 @@ private[zio] class OneShot[A] private (var value: A) {
    * Retrieves the value of the variable, blocking if necessary.
    */
   final def get(): A = {
-    if (value == null) throw new Error("Timed out waiting for variable to be set")
+    if (value == null) throw new Error("Cannot block for result to be set in Javascript")
     value
   }
 }
