+++ interop-cats/jvm/src/test/scala/scalaz/zio/interop/catzSpec.scala
@@ -17,7 +17,7 @@ import org.typelevel.discipline.Laws
 import org.typelevel.discipline.scalatest.Discipline
 import scalaz.zio.interop.catz._
 import cats.laws._
-import scalaz.zio.internal.Env
+import scalaz.zio.internal.impls.Env
 
 trait ConcurrentEffectLawsOverrides[F[_]] extends ConcurrentEffectLaws[F] {
